MSFN-master/requirements.txt
torch==1.6.0
torchvision==0.7.0
Pillow==8.1.1
pytorch-lightning==0.9.0
kornia==0.3.0
tqdm==4.54.0
yacs
albumentations
torch-dct
loguru
